+++ src/yaga/agent_review/__init__.py
@@ -3,16 +3,28 @@
 from yaga.agent_review.evaluation import LensOutcome, ReviewAggregate, evaluate_policy
 from yaga.agent_review.plan import ReviewPlan, ReviewPlanItem, build_plan, render_plan
 from yaga.agent_review.policy import AgentReviewPolicy, ReviewLens, load_policy
+from yaga.agent_review.results import (
+    AgentReviewResults,
+    LensResult,
+    evaluate_results,
+    load_results,
+    render_evaluation,
+)
 
 __all__ = [
     "AgentReviewPolicy",
+    "AgentReviewResults",
     "LensOutcome",
     "ReviewAggregate",
     "ReviewLens",
     "ReviewPlan",
     "ReviewPlanItem",
+    "LensResult",
     "build_plan",
     "evaluate_policy",
+    "evaluate_results",
     "load_policy",
+    "load_results",
     "render_plan",
+    "render_evaluation",
 ]